requirements.txt
asgiref==3.5.1
Django==3.2
django-crispy-forms==1.14.0
gunicorn==20.1.0
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
